:art: Move scrollable tab into new file
